Fix initCode
